Seeing if git doesn't lose my code like usual

+++ lib/models/profile/stock_chart.dart
@@ -20,7 +20,7 @@ class StockChart {
   factory StockChart.fromJson(Map<dynamic, dynamic> json) {
     return StockChart(
       date: json['date'],
-      close: json['close'],
+      close: json['close'].toDouble (),
       label: json['label'],
     );
   }
